Fix for adb mess up when saving to file in directory
When using
'adb backup -f "userApps/com.google.android.launcher.ab" -apk com.google.android.launcher.ab'
ADB messes up the backup. It creates a directory
"userApps/com.google.android.launcher.ab" instead of a file and then
complains that it cannot write to the file.

This is with ADB version 1.0.32 on a CM12.1 build on a Sony Xperia Z Ultra
and on Android 6.0.1 on a Nexus 7 (2013).

This fix cds to the userApps directory and then backups to a file without
subdirectory. This works fine.